+++ src/views/about_view.hpp
@@ -18,38 +18,16 @@
 
 #pragma once
 
-#include "rclone/rclone_manager.hpp"
-#include "daemon_proxy.hpp"
-#include "settings.hpp"
-#include "views/backends_view.hpp"
-#include "views/browser_view.hpp"
-#include "views/job_view.hpp"
-#include "views/about_view.hpp"
-#include "views/settings_view.hpp"
-#include <adwaita.h>
 #include <gtkmm.h>
 
 namespace saddle {
 
-class SaddleWindow : public Gtk::ApplicationWindow {
+class AboutView : public Gtk::Box {
 public:
-    explicit SaddleWindow(rclone::RcloneManager& manager, DaemonProxy* daemon_proxy,
-                          Settings& settings);
-
-    void show_toast(const char* message);
+    AboutView();
 
 private:
-    Settings&     m_settings;
-    DaemonProxy*  m_daemon_proxy  = nullptr;
-
-    AdwViewStack* m_view_stack    = nullptr;
-    Gtk::Widget*  m_toast_overlay = nullptr;
-
-    BackendsView  m_remotes_view;
-    JobView       m_job_view;
-    BrowserView   m_browser_view;
-    SettingsView  m_settings_view;
-    AboutView     m_about_view;
+    void setup_ui();
 };
 
 } // namespace saddle